Adds reader study create test
