+++ app/tests/reader_studies_tests/test_forms.py
@@ -1,7 +1,9 @@
 import pytest
 
+from grandchallenge.reader_studies.models import ReaderStudy
 from tests.factories import UserFactory
 from tests.reader_studies_tests.factories import ReaderStudyFactory
+from tests.reader_studies_tests.utils import get_rs_creator
 from tests.utils import get_view_for_user
 
 
@@ -53,3 +55,25 @@ def test_editor_update_form(client):
 
     rs.refresh_from_db()
     assert rs.readers_group.user_set.count() == 1
+
+
+@pytest.mark.django_db
+def test_reader_study_create(client):
+    # The study creator should automatically get added to the editors group
+    creator = get_rs_creator()
+
+    response = get_view_for_user(
+        viewname="reader-studies:create",
+        client=client,
+        method=client.post,
+        data={"title": "foo bar"},
+        follow=True,
+        user=creator,
+    )
+    assert response.status_code == 200
+
+    rs = ReaderStudy.objects.get(title="foo bar")
+
+    assert rs.slug == "foo-bar"
+    assert rs.is_editor(user=creator)
+    assert not rs.is_reader(user=creator)
